refactored companies file structure.
