+++ src/App.tsx
@@ -7,10 +7,10 @@ import UserRegistration from './components/UserRegistration';
 import { Route, Routes, Navigate } from 'react-router-dom';
 import Contacts from './pages/Contacts/Contacts';
 import NewContact from './pages/Contacts/components/NewContact';
-import Companies from './components/companies/Companies';
-import NewCompany from './components/companies/NewCompany';
+import Companies from './pages/Companies/Companies';
+import NewCompany from './pages/Companies/components/NewCompany';
 import NewJobApplication from './pages/JobApplications/components/NewJobApplication';
-import CompanyShow from './components/companies/CompanyShow';
+import CompanyShow from './pages/Companies/components/CompanyShow';
 import ApplicationsGrid from './pages/JobApplications/JobApplications';
 import ShowContact from './pages/Contacts/components/ShowContact';
 import JobApplication from './pages/JobApplications/components/showJobApplication';
